Now both server and client can send files
